refactor: Update component paths from /components into /src/components/
- Changed component import paths from "@/components" to "@/src/components" in components.json and playground.astro.
