+++ src/components/ui/image-card.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/src/lib/utils";
 
 type Props = {
   imageUrl: string;
