+++ src/pages/playground.astro
@@ -1,6 +1,6 @@
 ---
-import ImageCard from "@/components/ui/image-card";
-import Marquee from "@/components/ui/marquee";
+import ImageCard from "@/src/components/ui/image-card";
+import Marquee from "@/src/components/ui/marquee";
 
 const items = ["Item 1", "Item 2", "Item 3", "Item 4", "Item 5"];
